[ExposureNotification] Added support for multiple sources with ExposureNotificationServiceKeysDownloader
