+++ Modules/ExposureNotification/Backends/dp3t.py
@@ -3,16 +3,19 @@ from typing import *
 
 import pytz
 
-from Modules.ExposureNotification import exposure_notification_io
+from .base import BaseBackendKeysDownloader
 
 _dp3t_server_exposed_endpoint_path = "/v1/gaen/exposed/"
 
 _unix_epoch_datetime = datetime.datetime(year=1970, month=1, day=1, tzinfo=pytz.utc)
 
 
-class DP3TServiceKeysDownloader(exposure_notification_io.ExposureNotificationServiceKeysDownloader):
-    def generate_exposure_keys_export_endpoints_with_parameters(
-            self, server_endpoint_url: str, **kwargs) -> List[dict]:
+class DP3TBackendKeysDownloader(BaseBackendKeysDownloader):
+    def __init__(self, backend_identifier: str, server_endpoint_url: str):
+        super().__init__(backend_identifier=backend_identifier)
+        self.server_endpoint_url = server_endpoint_url
+
+    def generate_exposure_keys_export_endpoints_with_parameters(self, **kwargs) -> List[dict]:
         exposure_keys_export_endpoints = []
         dates = self.get_dates_from_parameters(**kwargs)
         for date in dates:
@@ -26,10 +29,10 @@ class DP3TServiceKeysDownloader(exposure_notification_io.ExposureNotificationSer
                 int((generation_datetime - _unix_epoch_datetime).total_seconds())
             endpoint_timestamp = str(generation_datetime_in_epoch_seconds) + "000"
             exposure_keys_export_endpoint = \
-                server_endpoint_url + _dp3t_server_exposed_endpoint_path + endpoint_timestamp
+                self.server_endpoint_url + _dp3t_server_exposed_endpoint_path + endpoint_timestamp
             exposure_keys_export_endpoints.append(dict(
                 endpoint=exposure_keys_export_endpoint,
                 generation_date=generation_date_string,
-                server_endpoint_url=server_endpoint_url,
+                server_endpoint_url=self.server_endpoint_url,
             ))
         return exposure_keys_export_endpoints
